# opt: minimize the number of buckets allocated when filtering histograms