#### opt: add new benchmark cases that filter histogram buckets

Release note: None

#### opt: minimize the number of buckets allocated when filtering histograms

When filtering a histogram, we previously allocated the same number of
buckets for the new histogram as the original histogram. Now we allocate
much fewer buckets in two ways:

  1. We allocate only the number of buckets including and after the
     first bucket that overlaps with the filtering span(s). For example,
     if the original histogram has 10 buckets and the first bucket that
     overlaps the span(s) is the bucket at index 3, then we allocate
     only 7+1 buckets (the additional bucket is the empty bucket
     representing the lower bound of the filtered histogram).

  2. In the special case where we have a single filtering span that
     overlaps a single bucket, we allocate just 2 buckets. This case is
     common for equality filters and small range filters.

Informs #136002

Release note: None