@mgartner 2024-12-07
This reduces the number of allocated bytes, not allocations. So the overall performance improvement may be small or zero. Nevertheless, I think it's generally a good thing to avoid allocating 8kB (`200 buckets * 40 bytes per bucket`) in the common case when there is no need to.

Optimizer micro benchmarks show the improvement in allocated bytes:

```
name                                                                               old time/op    new time/op    delta
Phases/single-col-histogram-range/Simple/Parse                                       5.35µs ± 1%    5.40µs ± 1%   +0.99%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm                              35.2µs ± 0%    35.1µs ± 0%     ~     (p=0.690 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNorm                                53.7µs ± 1%    53.6µs ± 2%     ~     (p=0.151 n=5+5)
Phases/single-col-histogram-range/Simple/Explore                                     59.5µs ± 3%    60.1µs ± 6%     ~     (p=1.000 n=5+5)
Phases/single-col-histogram-range/Simple/ExecBuild                                   62.5µs ± 0%    62.4µs ± 0%     ~     (p=0.310 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm                  22.9µs ± 0%    22.9µs ± 0%     ~     (p=0.056 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm                    22.9µs ± 0%    22.9µs ± 0%     ~     (p=0.222 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore                                   25.7µs ± 1%    25.9µs ± 0%     ~     (p=0.690 n=5+5)
Phases/single-col-histogram-range/Prepared/ExecBuild                                 27.9µs ± 1%    27.9µs ± 1%     ~     (p=0.310 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/Parse                         8.05µs ± 0%    8.09µs ± 0%   +0.47%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/OptBuildNoNorm                26.8µs ± 0%    26.2µs ± 1%   -2.26%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/OptBuildNorm                  34.2µs ± 1%    32.4µs ± 0%   -5.35%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/Explore                       40.5µs ± 0%    38.9µs ± 1%   -3.83%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/ExecBuild                     44.9µs ± 0%    42.6µs ± 0%   -5.06%  (p=0.016 n=5+4)
Phases/single-col-histogram-bounded-range-small/Prepared/AssignPlaceholdersNorm      10.3µs ± 1%     9.4µs ± 1%   -9.43%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Prepared/Explore                     14.0µs ± 1%    13.2µs ± 1%   -5.57%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Prepared/ExecBuild                   16.2µs ± 1%    15.4µs ± 0%   -4.73%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/Parse                           8.06µs ± 0%    8.11µs ± 0%   +0.58%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNoNorm                  30.7µs ± 0%    30.7µs ± 1%     ~     (p=0.690 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNorm                    42.6µs ± 0%    41.7µs ± 0%   -1.97%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/Explore                         48.8µs ± 0%    48.2µs ± 0%   -1.16%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/ExecBuild                       53.2µs ± 0%    52.3µs ± 0%   -1.66%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNoNorm      18.4µs ± 0%    17.7µs ± 0%   -3.83%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNorm        14.6µs ± 0%    14.4µs ± 1%   -1.42%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/Explore                       18.4µs ± 1%    18.4µs ± 0%     ~     (p=1.000 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/ExecBuild                     20.6µs ± 0%    20.4µs ± 1%   -0.72%  (p=0.016 n=5+5)

name                                                                               old alloc/op   new alloc/op   delta
Phases/single-col-histogram-range/Simple/Parse                                       3.40kB ± 0%    3.40kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm                              15.6kB ± 0%    15.6kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNorm                                20.6kB ± 0%    20.6kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/Explore                                     23.6kB ± 0%    23.6kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/ExecBuild                                   25.3kB ± 0%    25.3kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm                  7.37kB ± 0%    7.37kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm                    7.37kB ± 0%    7.37kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore                                   9.79kB ± 0%    9.79kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/ExecBuild                                 11.4kB ± 0%    11.4kB ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/Parse                         3.64kB ± 0%    3.64kB ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/OptBuildNoNorm                18.2kB ± 0%    14.2kB ± 0%  -22.01%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/OptBuildNorm                  25.6kB ± 0%    17.6kB ± 0%  -31.37%  (p=0.000 n=5+4)
Phases/single-col-histogram-bounded-range-small/Simple/Explore                       28.1kB ± 0%    20.0kB ± 0%  -28.63%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/ExecBuild                     29.7kB ± 0%    21.7kB ± 0%  -27.04%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Simple/ExecBuild                     29.7kB ± 0%    21.7kB ± 0%  -27.04%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Prepared/AssignPlaceholdersNoNorm    10.4kB ± 0%     7.4kB ± 0%  -28.42%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Prepared/AssignPlaceholdersNorm      9.76kB ± 0%    5.75kB ± 0%  -41.14%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-small/Prepared/Explore                     12.2kB ± 0%     8.2kB ± 0%  -32.95%  (p=0.000 n=5+4)
Phases/single-col-histogram-bounded-range-small/Prepared/ExecBuild                   13.8kB ± 0%     9.8kB ± 0%  -29.04%  (p=0.029 n=4+4)
Phases/single-col-histogram-bounded-range-big/Simple/Parse                           3.64kB ± 0%    3.64kB ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNoNorm                  18.9kB ± 0%    15.9kB ± 0%  -15.59%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNorm                    26.9kB ± 0%    21.0kB ± 0%  -21.90%  (p=0.000 n=5+4)
Phases/single-col-histogram-bounded-range-big/Simple/Explore                         29.3kB ± 0%    23.5kB ± 0%  -20.07%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Simple/ExecBuild                       31.0kB ± 0%    25.1kB ± 0%  -19.00%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNoNorm      10.3kB ± 0%     7.4kB ± 0%  -28.46%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNorm        10.4kB ± 0%     7.5kB ± 0%  -28.30%  (p=0.008 n=5+5)
Phases/single-col-histogram-bounded-range-big/Prepared/Explore                       12.8kB ± 0%     9.9kB ± 0%     ~     (p=0.079 n=4+5)
Phases/single-col-histogram-bounded-range-big/Prepared/ExecBuild                     14.5kB ± 0%    11.5kB ± 0%  -20.35%  (p=0.000 n=5+4)

name                                                                               old allocs/op  new allocs/op  delta
Phases/single-col-histogram-range/Simple/Parse                                         21.0 ± 0%      21.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm                                79.0 ± 0%      79.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm                                79.0 ± 0%      79.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNorm                                  84.0 ± 0%      84.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/Explore                                       96.0 ± 0%      96.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/ExecBuild                                      106 ± 0%       106 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm                    26.0 ± 0%      26.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm                      26.0 ± 0%      26.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore                                     37.0 ± 0%      37.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/ExecBuild                                   47.0 ± 0%      47.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/Parse                           25.0 ± 0%      25.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/OptBuildNorm                     148 ± 0%       148 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/Explore                          159 ± 0%       159 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Simple/ExecBuild                        169 ± 0%       169 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Prepared/AssignPlaceholdersNoNorm      69.0 ± 0%      69.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Prepared/AssignPlaceholdersNorm        53.0 ± 0%      53.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Prepared/Explore                       64.0 ± 0%      64.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-small/Prepared/ExecBuild                     74.0 ± 0%      74.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/Parse                             25.0 ± 0%      25.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNoNorm                     132 ± 0%       132 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/OptBuildNorm                       182 ± 0%       182 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/Explore                            193 ± 0%       193 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Simple/ExecBuild                          203 ± 0%       203 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNoNorm        67.0 ± 0%      67.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Prepared/AssignPlaceholdersNorm          70.0 ± 0%      70.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Prepared/Explore                         81.0 ± 0%      81.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Prepared/ExecBuild                       91.0 ± 0%      91.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-bounded-range-big/Prepared/ExecBuild                       91.0 ± 0%      91.0 ± 0%     ~     (all equal)
```

End-to-end benchmarks show the improvement too, though I think the latency decrease measured here was probably just noise:

```
name                                                                      old time/op    new time/op    delta
EndToEnd/single-col-histogram-range/vectorize=on/Simple                      248µs ± 1%     242µs ± 0%   -2.58%  (p=0.000 n=9+8)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared                    152µs ± 2%     147µs ± 2%   -3.40%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-range/vectorize=off/Simple                     244µs ± 1%     237µs ± 2%   -3.10%  (p=0.000 n=7+10)
EndToEnd/single-col-histogram-range/vectorize=off/Prepared                   149µs ± 3%     144µs ± 1%   -3.68%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Simple        255µs ± 5%     246µs ± 3%   -3.78%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared      177µs ± 3%     170µs ± 2%   -3.70%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Simple       261µs ± 3%     253µs ± 2%   -3.10%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Prepared     176µs ± 1%     172µs ± 2%   -2.64%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Simple          290µs ± 1%     287µs ± 2%   -1.00%  (p=0.034 n=8+10)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared        199µs ± 0%     191µs ± 1%   -4.16%  (p=0.000 n=9+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Simple         293µs ± 0%     289µs ± 0%   -1.37%  (p=0.001 n=7+7)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Prepared       192µs ± 1%     186µs ± 1%   -3.29%  (p=0.000 n=8+8)

name                                                                      old alloc/op   new alloc/op   delta
EndToEnd/single-col-histogram-range/vectorize=on/Simple                     81.9kB ± 1%    81.9kB ± 1%     ~     (p=0.968 n=10+9)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared                   36.4kB ± 2%    36.8kB ± 6%     ~     (p=1.000 n=9+10)
EndToEnd/single-col-histogram-range/vectorize=off/Simple                    41.5kB ± 0%    42.0kB ± 6%     ~     (p=0.815 n=8+9)
EndToEnd/single-col-histogram-range/vectorize=off/Prepared                  38.2kB ± 0%    38.2kB ± 0%     ~     (p=0.340 n=8+8)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Simple       43.7kB ± 1%    39.5kB ± 0%   -9.54%  (p=0.000 n=8+9)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared     32.6kB ± 0%    28.6kB ± 0%  -12.32%  (p=0.000 n=9+9)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Simple      45.1kB ± 0%    41.1kB ± 0%   -8.79%  (p=0.000 n=10+10)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Prepared    34.9kB ±11%    29.8kB ± 0%  -14.60%  (p=0.000 n=10+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Simple         85.3kB ± 0%    82.3kB ± 0%   -3.52%  (p=0.000 n=8+8)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared       74.2kB ± 0%    71.3kB ± 0%   -4.01%  (p=0.000 n=9+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Simple        45.6kB ± 0%    42.7kB ± 0%   -6.36%  (p=0.000 n=9+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Prepared      34.4kB ± 0%    31.5kB ± 0%   -8.47%  (p=0.000 n=8+8)

name                                                                      old allocs/op  new allocs/op  delta
EndToEnd/single-col-histogram-range/vectorize=on/Simple                        317 ± 0%       320 ± 3%     ~     (p=0.247 n=8+9)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared                      308 ± 0%       308 ± 0%     ~     (p=0.165 n=8+9)
EndToEnd/single-col-histogram-range/vectorize=off/Simple                       311 ± 0%       311 ± 0%     ~     (p=1.000 n=8+8)
EndToEnd/single-col-histogram-range/vectorize=off/Prepared                     289 ± 0%       289 ± 0%     ~     (all equal)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Simple          350 ± 0%       350 ± 0%     ~     (p=0.200 n=8+8)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared        267 ± 0%       267 ± 0%     ~     (all equal)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Simple         345 ± 0%       345 ± 0%     ~     (p=0.059 n=10+8)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=off/Prepared       261 ± 0%       261 ± 0%     ~     (p=0.941 n=8+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Simple            368 ± 0%       367 ± 0%     ~     (p=0.557 n=8+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared          284 ± 0%       284 ± 0%     ~     (all equal)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Simple           361 ± 0%       362 ± 0%     ~     (p=0.347 n=9+9)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=off/Prepared         278 ± 0%       278 ± 0%     ~     (p=0.138 n=9+7)
```